chore: add new 'onchain-reconcile' module to facade
