+++ core/api/src/services/ledger/facade/index.ts
@@ -2,6 +2,7 @@ export * from "./intraledger"
 export * from "./offchain-receive"
 export * from "./offchain-send"
 export * from "./onchain-receive"
+export * from "./onchain-reconcile"
 export * from "./onchain-send"
 export * from "./reconciliation"
 export * from "./static-account-ids"
